% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demrit.R
\name{demerchart}
\alias{demerchart}
\title{Demerit control chart using fuzzy cmeans
demrchart function}
\usage{
demerchart(
  x,
  fcm.res,
  df,
  z,
  b1,
  b2,
  b3,
  b4,
  q,
  dn,
  CL,
  UCL,
  LCL,
  n,
  averagenumber_demerits,
  control_chart
)
}
\arguments{
\item{x}{dataset used}

\item{fcm.res}{returns the fuzzy clustering process}

\item{df}{membership value sum per class}

\item{z}{mean of demerit control chart}

\item{b1}{membership value with number of defects for class A}

\item{b2}{membership value with number of defects for class B}

\item{b3}{membership value with number of defects for class C}

\item{b4}{membership value with number of defects for class D}

\item{q}{standard deviation for the demerit control chart}

\item{dn}{summation of membership value per column}

\item{CL}{center line}

\item{UCL}{upper control limit}

\item{LCL}{lower control limit}

\item{n}{total number of sample size}

\item{averagenumber_demerits}{average}

\item{control_chart}{contro chart}
}
\value{
limits
}
\description{
Demerit control chart using fuzzy cmeans
demrchart function
}
\examples{
demerchart(orangejuice,fcm.res,n=2700)
}
